reworked mapping and added constraint

+++ Core/Inc/dynamixel_xl320.h
@@ -68,7 +68,7 @@ void XL320_set_led_ON(UART_HandleTypeDef *m_huart, uint8_t color);
 
 void XL320_set_led_OFF(UART_HandleTypeDef *m_huart);
 
-void XL320_set_pos(UART_HandleTypeDef *m_huart, uint8_t pos);
+void XL320_set_pos(UART_HandleTypeDef *m_huart, uint16_t pos);
 
 uint16_t XL320_read_load(UART_HandleTypeDef *m_huart);
